moved more actions into functions , fixed missing presets and feedbacks. Linted all.
