+++ package.json
@@ -34,6 +34,7 @@
 	},
 	"devDependencies": {
 		"@companion-module/tools": "^2.1.0",
+		"eslint": "9.14.0",
 		"husky": "^9.1.6",
 		"prettier": "^3.3.3",
 		"typedoc": "^0.26.8",
