+++ src/actions/ActionId.ts
@@ -1,22 +1,4 @@
 export enum ActionId {
-	
-	
-	//Super Source
-	SuperSourceEnable = 'superSourceEnable',
-	SuperSourceSource1 = 'superSourceSource1',
-	SuperSourceSource2 = 'superSourceSource2',
-	SuperSourceBackground = 'superSourceBackground',
-	SuperSourceControlStyle = 'superSourceControlStyle',
-	SuperSourceControlYPosition = 'superSourceControlYPosition',
-	SuperSourceMaskEnable = 'superSourceMaskEnable',
-	SuperSourceMaskHStart = 'superSourceMaskHStart',
-	SuperSourceMaskVStart = 'superSourceMaskVStart',
-	SuperSourceMaskHEnd = 'superSourceMaskHEnd',
-	SuperSourceMaskVEnd = 'superSourceMaskVEnd',
-	SuperSourceBorderWidth = 'superSourceBorderWidth',
-	SuperSourceBorderHue = 'superSourceBorderHue',
-	SuperSourceBorderSaturation = 'superSourceBorderSaturation',
-	SuperSourceBorderBrightness = 'superSourceBorderBrightness',
 	//UpStream Key
 	KeyOnAir = 'keyOnAir',
 	TransitionSource = 'transitionSource',
@@ -85,41 +67,4 @@ export enum ActionId {
 	PipBorderHue = 'pipBorderHue',
 	PipBorderSaturation = 'pipBorderSaturation',
 	PipBorderBrightness = 'pipBorderBrightness',
-
-	//Macro
-	MacroRecord = 'macroRecord',
-	MacroInfo = 'macroInfo',
-	GetMacroInfoAll = 'getMacroInfoAll',
-	MacroRun = 'macroRun',
-	RemoveMacro = 'removeMacro',
-	MacroSleep = 'macroSleep',
-	//Settings
-	SrcName = 'srcName',
-	MvMeter = 'mvMeter',
-	MvLayout = 'mvLayout',
-	InputWindowLayout = 'inputWindowLayout',
-	Marker = 'marker',
-	MicInput = 'micInput',
-	RecordFileName = 'recordFileName',
-	SrcSelection = 'srcSelection',
-	AuxSource = 'auxSource',
-	OutFormat = 'outFormat',
-	OutputColorSpace = 'outputColorSpace',
-	OutSource = 'outSource',
-	NetworkProtocol = 'networkProtocol',
-	NetworkIpAddress = 'networkIpAddress',
-	NetworkSubnetMask = 'networkSubnetMask',
-	NetworkGateway = 'networkGateway',
-	NetworkPrimaryDNS = 'networkPrimaryDNS',
-	NetworkSecondaryDNS = 'networkSecondaryDNS',
-	Quality = 'quality',
-	Panel = 'panel',
-	SdFormat = 'sdFormat',
-	SystemReset = 'systemReset',
-	RecordStatus = 'recordStatus',
-	RecordTime = 'recordTime',
-	SdCardStatus = 'sdCardStatus',
-	SdFree = 'sdFree',
-	LiveStatus = 'liveStatus',
-	LiveInfo = 'liveInfo',
 }
