+++ src/functions/macro/actionId.ts
@@ -1,0 +1,8 @@
+export enum ActionId {
+	MacroRecord = 'macroRecord',
+	MacroInfo = 'macroInfo',
+	GetMacroInfoAll = 'getMacroInfoAll',
+	MacroRun = 'macroRun',
+	RemoveMacro = 'removeMacro',
+	MacroSleep = 'macroSleep',
+}
