+++ src/functions/macro/feedbackId.ts
@@ -1,0 +1,3 @@
+export enum FeedbackId {
+	Macro = 'macro',
+}
