+++ src/functions/playback/feedbackId.ts
@@ -1,0 +1,7 @@
+export enum FeedbackId {
+	PlaybackMode = 'playbackMode',
+	PlaybackRepeat = 'playbackRepeat',
+	PlaybackPause = 'playbackPause',
+	PlaybackBar = 'playbackBar',
+	PlayFile = 'playFile',
+}
