+++ src/functions/playback/index.ts
@@ -1,2 +1,4 @@
 export * as PlaybackActions from './actions'
 export * as PlaybackState from './state'
+export * as PlaybackFeedbacks from './feedbacks'
+export * as PlaybackPresets from './presets'
