+++ src/functions/settings/actionId.ts
@@ -1,0 +1,30 @@
+export enum ActionId {
+	SrcName = 'srcName',
+	MvMeter = 'mvMeter',
+	MvLayout = 'mvLayout',
+	InputWindowLayout = 'inputWindowLayout',
+	Marker = 'marker',
+	MicInput = 'micInput',
+	RecordFileName = 'recordFileName',
+	SrcSelection = 'srcSelection',
+	AuxSource = 'auxSource',
+	OutFormat = 'outFormat',
+	OutputColorSpace = 'outputColorSpace',
+	OutSource = 'outSource',
+	NetworkProtocol = 'networkProtocol',
+	NetworkIpAddress = 'networkIpAddress',
+	NetworkSubnetMask = 'networkSubnetMask',
+	NetworkGateway = 'networkGateway',
+	NetworkPrimaryDNS = 'networkPrimaryDNS',
+	NetworkSecondaryDNS = 'networkSecondaryDNS',
+	Quality = 'quality',
+	Panel = 'panel',
+	SdFormat = 'sdFormat',
+	SystemReset = 'systemReset',
+	RecordStatus = 'recordStatus',
+	RecordTime = 'recordTime',
+	SdCardStatus = 'sdCardStatus',
+	SdFree = 'sdFree',
+	LiveStatus = 'liveStatus',
+	LiveInfo = 'liveInfo',
+}
