+++ src/functions/settings/feedbackId.ts
@@ -1,0 +1,10 @@
+export enum FeedbackId {
+	InputWindowLayout = 'inputWindowLayout',
+	SrcSelection = 'srcSelection',
+	OutputColorSpace = 'outputColorSpace',
+	OutputFormat = 'outFormat',
+	MicInput = 'micInput',
+	MvLayout = 'mvLayout',
+	SettingOutSource = 'settingOutSource',
+	AuxBG = 'auxBG',
+}
