+++ src/functions/settings/index.ts
@@ -1,0 +1,3 @@
+export * as SettingsActions from './actions'
+export * as SettingsState from './state'
+export * as SettingsFeedbacks from './feedbacks'
