+++ src/functions/superSource/index.ts
@@ -1,3 +1,4 @@
 export * as SuperSourceActions from './actions'
 export * as SuperSourceFeedbacks from './feedbacks'
 export * as SuperSourceState from './state'
+export * as SuperSourcePresets from './presets'
