+++ src/state.ts
@@ -1,4 +1,3 @@
-import { Choice } from './choices'
 import { MixEffectState } from './functions/mixEffect'
 import { StreamingState } from './functions/streaming'
 import { LiveState } from './functions/live'
@@ -8,6 +7,8 @@ import { RecordState } from './functions/record'
 import { SuperSourceState } from './functions/superSource'
 import { AudioMixerState } from './functions/audioMixer'
 import { DownstreamKeyerState } from './functions/downstreamKeyer'
+import { SettingsState } from './functions/settings'
+import { MacroState } from './functions/macro'
 
 type GoStreamState = {
 	MixEffect: MixEffectState.State
@@ -19,6 +20,8 @@ type GoStreamState = {
 	SuperSource: SuperSourceState.State
 	AudioMixer: AudioMixerState.State
 	DownstreamKeyer: DownstreamKeyerState.State
+	Settings: SettingsState.State
+	Macro: MacroState.State
 	infos: {
 		protocolVersion: string
 		deviceType: number
@@ -38,28 +41,6 @@ type GoStreamState = {
 		UpStreamKeyType: number
 		ArrayKeySourceFill: number[]
 	}
-	SettingsProp: {
-		AuxSource: number
-		OutSource: {
-			hdmi1: Choice
-			hdmi2: Choice
-			uvc: Choice
-		}
-		SettingsInputWindowLayout: number
-		MvMeter: number[]
-		SourceSelection: number[]
-		OutputColorSpace: number[]
-		OutputFormat: number
-		MicInput: number[]
-		MvLayout: number
-	}
-	StillProp: {
-		Still1: number
-		Still2: number
-	}
-	MacroProp: {
-		macroProperties: number[]
-	}
 }
 
 export function Create(): GoStreamState {
@@ -73,6 +54,8 @@ export function Create(): GoStreamState {
 		...SuperSourceState.create(),
 		...AudioMixerState.create(),
 		...DownstreamKeyerState.create(),
+		...SettingsState.create(),
+		...MacroState.create(),
 		infos: {
 			protocolVersion: '1.0',
 			deviceType: 0,
@@ -91,28 +74,6 @@ export function Create(): GoStreamState {
 			OnAir: false,
 			UpStreamKeyType: 0,
 			ArrayKeySourceFill: [0, 0, 0, 0],
-		},
-		SettingsProp: {
-			AuxSource: 0,
-			OutSource: {
-				hdmi1: { id: 0, label: 'Input1' },
-				hdmi2: { id: 0, label: 'Input1' },
-				uvc: { id: 0, label: 'Input1' },
-			},
-			SettingsInputWindowLayout: 0,
-			MvMeter: [0, 0, 0, 0, 0, 0],
-			SourceSelection: [0, 0, 0, 0], // INPUT1-4
-			OutputColorSpace: [0, 0],
-			OutputFormat: 0,
-			MicInput: [0, 0],
-			MvLayout: 0,
-		},
-		StillProp: {
-			Still1: 0,
-			Still2: 0,
-		},
-		MacroProp: {
-			macroProperties: [],
 		},
 	}
 }
